+++ Core/Headers/core/math/bigMatrix.h
@@ -6,20 +6,20 @@ struct _bigMatrix
 {
 	array(float) Values;
 	// The Height of the matrix
-	size_t Rows;
+	ulong Rows;
 	// The Width of the matrix
-	size_t Columns;
+	ulong Columns;
 };
 
 typedef struct _bigMatrix* BigMatrix;
 
 extern const struct _arbitraryMatrixMethods
 {
-	BigMatrix(*Create)(size_t rows, size_t columns);
+	BigMatrix(*Create)(ulong rows, ulong columns);
 	// Multiplies the big matrix with the given vector and appends the results to the destinationVector
 	void (*MultiplyVector)(BigMatrix, array(float) vector, array(float) destinationVector);
-	void (*Resize)(BigMatrix, size_t rows, size_t columns);
-	float* (*At)(BigMatrix, size_t row, size_t column);
+	void (*Resize)(BigMatrix, ulong rows, ulong columns);
+	float* (*At)(BigMatrix, ulong row, ulong column);
 	void (*Clear)(BigMatrix);
 	void (*Dispose)(BigMatrix);
 } BigMatrices;
