+++ Core/Headers/core/math/floats.h
@@ -3,7 +3,7 @@
 #include "core/file.h"
 
 struct _floatMethods {
-	bool (*TryDeserialize)(const char* buffer, size_t bufferLength, float* out_float);
+	bool (*TryDeserialize)(const char* buffer, ulong bufferLength, float* out_float);
 	void(*SerializeStream)(File stream, float value);
 };
